requirements.txt
python-telegram-bot==13.11
environs==9.5.0
google-cloud-dialogflow==2.13.0
requests==2.27.1
vk_api==11.9.7